feat: Consolidate getData
